requirements.txt
addict==2.4.0
attrs==23.1.0
backcall==0.2.0
bagpy==0.5
bitarray==2.8.3
bitstring==4.1.2
catkin-pkg==1.0.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.2.0
colorama==0.4.6
cPython==0.0.6
cycler==0.11.0
Cython==3.0.2
decorator==5.1.1
dnspython==2.3.0
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
genmsg @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/genmsg-0.5.12-py2.py3-none-any.whl
genpy @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/genpy-0.6.8-py2.py3-none-any.whl
gnupg==2.3.1
h5py==3.8.0
hdf5plugin==4.3.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==7.34.0
jedi==0.19.1
Jinja2==3.0.3
joblib==1.3.2
JsonForm==0.0.2
jsonschema==4.17.3
JsonSir==0.0.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
MultiScaleDeformableAttention==1.0
numpy==1.21.6
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.1
pandas==1.3.5
parso==0.8.3
peppercorn==0.6
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.40
psutil==5.9.6
py3rosmsgs==1.18.2
pycocotools-windows @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/pycocotools_windows-2.0.0.2-cp37-cp37m-win_amd64.whl
pycryptodome==3.19.0
pycryptodomex==3.19.0
Pygments==2.16.1
pymongo==4.5.0
pyparsing==3.1.1
pyrsistent==0.19.3
pyserial==3.5
pytest==7.4.3
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
python-resources==0.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
Resource==0.2.1
rosbag @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/rosbag-1.15.11-py2.py3-none-any.whl
roscpp @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/roscpp-1.15.7-py2.py3-none-any.whl
rosgraph @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/rosgraph-1.15.11-py2.py3-none-any.whl
rosgraph-msgs==0.0.1
roslib @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/roslib-1.14.6-py2.py3-none-any.whl
roslz4 @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/roslz4-1.14.3.post2-cp37-cp37m-win_amd64.whl
rospkg==1.5.0
rospy @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/rospy-1.15.11-py2.py3-none-any.whl
safetensors==0.3.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
setuptools-scm==7.1.0
six==1.16.0
sklearn==0.0.post9
std-msgs==0.0.1
termcolor==2.3.0
threadpoolctl==3.1.0
timm==0.9.7
tomli==2.0.1
torch @ file:///D:/ANACONDA/envs/dino/Lib/site-packages/torch-1.9.1%2Bcu111-cp37-cp37m-win_amd64.whl
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.10
wincertstore==0.2
yapf==0.40.1
zipp==3.15.0
